+++ src/Search/SearchServiceProvider.php
@@ -14,20 +14,19 @@ class SearchServiceProvider extends ServiceProvider
 
     public function boot()
     {
-        config()->set('scout.driver', 'tntsearch');
         config()->set('scout.queue', true);
         config()->set('scout.soft_delete', true);
         config()->set('scout.tntsearch', [
             'storage'  => storage_path(), //place where the index files will be stored
-            'fuzziness' => env('TNTSEARCH_FUZZINESS', false),
+            'fuzziness' => false,
             'fuzzy' => [
                 'prefix_length' => 2,
                 'max_expansions' => 50,
                 'distance' => 2
             ],
             'asYouType' => false,
-            'searchBoolean' => env('TNTSEARCH_BOOLEAN', false),
-            'maxDocs' => env('TNTSEARCH_MAX_DOCS', 500),
+            'searchBoolean' => false,
+            'maxDocs' => 500,
         ]);
     }